### 2024-01-12 #6223 @editorialbot
[PRE REVIEW]: statemodify: a Python framework to facilitate accessible exploratory modeling for discovering drought vulnerabilities
**Submitting author:** @rg727 (<a href="http://orcid.org/0000-0003-3538-0675">Rohini Gupta</a>)
**Repository:** https://github.com/IMMM-SFA/statemodify
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.4
