### 2024-01-12 5278060
fix paper doi refs
