+++ paper/paper.bib
@@ -26,6 +26,7 @@
   author={Salehabadi, Homa and Tarboton, David and Kuhn, Eric and Udall, Brad and Wheeler, Kevin and Rosenberg, David and Goeking, Sara and Schmidt, John C},
   journal={The Future of the Colorado River Project},
   year={2020},
+  doi={10.4211/hs.d3efcf0c930646fd9ef4f17c56436d20},
   publisher={Utah State University, Quinney College of Natural Resources Center for~�}
 }
 
@@ -128,7 +129,7 @@
   title={Decision making under deep uncertainty: from theory to practice},
   author={Marchau, Vincent AWJ and Walker, Warren E and Bloemen, Pieter JTM and Popper, Steven W},
   year={2019},
-  doi = {10.1287/opre.41.3.435},
+  doi = {10.1007/978-3-030-05252-2},
   publisher={Springer Nature}
 }
 
@@ -174,7 +175,7 @@
   number={6},
   pages={480--489},
   year={2004},
-  doi = {10.1061/(ASCE)0733-9496(2004)130:6(480},
+  doi = {10.1061/(ASCE)0733-9496(2004)130:6(480)},
   publisher={American Society of Civil Engineers}
 }
 
