### 2024-01-31 #6223 @ekblad
Hi, this looks good, I will take a look at the checklist and repo in the
next couple weeks.

On Mon, Jan 29, 2024 at 9:29 AM Taher Chegini ***@***.***>
wrote:

> Hi @barneydobson <https://github.com/barneydobson>, @ekblad
> <https://github.com/ekblad>, and @pfeiffea <https://github.com/pfeiffea>!
> Would you like to review this submission to the Journal for Open Source
> Software? Our reviews are checklist-driven and openly conducted on GitHub
> over a timeline of 4–6 weeks. Because the process is much more iterative
> and interactive than a traditional paper review, we would ask you to start
> within the next 2 weeks. Here are reviewer guidelines for reference:
> joss.readthedocs.io/en/latest/reviewer_guidelines.html
> Thanks for your consideration.
>
