### 2024-01-31 #6223 @cheginit
Hi @gutabeshu! Would you like to review this submission to the Journal for Open Source Software? Our reviews are checklist-driven and openly conducted on GitHub over a timeline of 4–6 weeks. Because the process is much more iterative and interactive than a traditional paper review, we would ask you to start within the next 2 weeks. Here are reviewer guidelines for reference: [joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
Thanks for your consideration.
