### 2024-02-05 #6325 @editorialbot
[REVIEW]: statemodify: a Python framework to facilitate accessible exploratory modeling for discovering drought vulnerabilities
**Submitting author:** @rg727 (<a href="http://orcid.org/0000-0003-3538-0675">Rohini Gupta</a>)
**Repository:** https://github.com/IMMM-SFA/statemodify
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.1
