### 2024-02-06 #6325 @rg727
@barneydobson, @ekblad thanks so much for agreeing to review our package. We are super excited to get some outside feedback. In case it's not clear, if you follow the links in the GitHub repo to the 5 tutorial Jupyter Notebooks, that's the easiest way to start testing out functionality.
