+++ .github/workflows/build.yml
@@ -8,7 +8,7 @@ jobs:
 
       strategy:
         matrix:
-          os: [ubuntu-latest]
+          os: [ubuntu-latest, windows-latest]
 
       env:
         OS: ${{ matrix.os }}
