### 2024-02-20 715427b
add pre-commit to deps
