+++ statemodify/tests/test_modify.py
@@ -14,11 +14,11 @@ class TestModify(unittest.TestCase):
     COMP_LEFT = "test    "
     COMP_NONE = "test"
     COMP_ZEROS = "0.10"
-    COMP_DATA = {'a': [0], 'b': ['alpha  '], 'c': [2.0]}
+    COMP_DATA = {"a": [0], "b": ["alpha  "], "c": [2.0]}
     COMP_FILE = Path("/my/output/directory") / "file_S0_test.txt"
     COMP_STR = "item   focus\n"
-    COMP_ADJ_DF_ADD = pd.DataFrame({"a": [10., 20.1], "b": [-1., -1.9]})
-    COMP_ADJ_DF_MULTIPLY = pd.DataFrame({"a": [10., 2.], "b": [-1., -0.2]})
+    COMP_ADJ_DF_ADD = pd.DataFrame({"a": [10.0, 20.1], "b": [-1.0, -1.9]})
+    COMP_ADJ_DF_MULTIPLY = pd.DataFrame({"a": [10.0, 2.0], "b": [-1.0, -0.2]})
     FIELD_DICT = {"a": [], "b": [], "c": []}
     COLUMN_WIDTHS = {"a": 1, "b": 7, "c": 3}
     COLUMN_LIST = ["a", "b", "c"]
@@ -32,38 +32,28 @@ class TestModify(unittest.TestCase):
 
         # ensure error is raised when value other than right or left is passed
         with self.assertRaises(AssertionError):
-            s = modify.set_alignment(value="test",
-                                     n_spaces=0,
-                                     align="center")
+            s = modify.set_alignment(value="test", n_spaces=0, align="center")
 
         # check for left padding (right alignment) of four on value; total length of eight
-        s = modify.set_alignment(value="test",
-                                 n_spaces=4,
-                                 align="right")
+        s = modify.set_alignment(value="test", n_spaces=4, align="right")
 
         self.assertEqual(self.COMP_RIGHT, s)
         self.assertEqual(len(self.COMP_RIGHT), len(s))
 
         # check for right padding (left alignment) of four on value; total length of eight
-        s = modify.set_alignment(value="test",
-                                 n_spaces=4,
-                                 align="left")
+        s = modify.set_alignment(value="test", n_spaces=4, align="left")
 
         self.assertEqual(self.COMP_LEFT, s)
         self.assertEqual(len(self.COMP_LEFT), len(s))
 
         # check for no space of four on value; total length of eight
-        s = modify.set_alignment(value="test",
-                                 n_spaces=0,
-                                 align="left")
+        s = modify.set_alignment(value="test", n_spaces=0, align="left")
 
         self.assertEqual(self.COMP_NONE, s)
         self.assertEqual(len(self.COMP_NONE), len(s))
 
         # check for n_spaces smaller than length of value of four; total length of eight
-        s = modify.set_alignment(value="test",
-                                 n_spaces=-2,
-                                 align="left")
+        s = modify.set_alignment(value="test", n_spaces=-2, align="left")
 
         self.assertEqual(self.COMP_NONE, s)
         self.assertEqual(len(self.COMP_NONE), len(s))
@@ -73,16 +63,17 @@ class TestModify(unittest.TestCase):
 
         # ensure error is raised when value length is greater than expected width
         with self.assertRaises(TypeError):
-            s = modify.pad_with_spaces(value="  test  ",
-                                       expected_width=2,
-                                       align="right")
+            s = modify.pad_with_spaces(
+                value="  test  ", expected_width=2, align="right"
+            )
+
+            del s
 
     def test_add_zero_padding(self):
         """Tests for the add_zero_padding function."""
 
[…]
