+++ .github/workflows/draft-pdf.yml
@@ -1,7 +1,7 @@
 ---
-on:
+on:    # yamllint disable-line rule:truthy
   push:
-    branches:
+    branches-ignore:
       - '**'
 jobs:
   paper:
