### 2024-02-26 IMMM-SFA/statemodify#54 @ekblad
notebook UI
re: openjournals/joss-reviews#6325
really like the jupyter execution and talk-through with this library. possible improvements:

1. when the user should see something, show it to them. example is when saying "In the input_files directory, you can open the sj2015B_S0_1.ddr file and see that the Admin # of our selected user has now become 1.0000." two ideas for this:
1a. use pandas immediately after to query the information that was described in words
1b. print directory contents or trees from cells using `!`
2. when executing model runs, suppress model output to avoid bogging down the instance.
3. in structuring notebooks as the application of two or more major functions, minimize the surrounding code (do users need to cast several dictionary elements to float?).
4. collapsing config into a single .json or .yaml would save some lines (do users really need to set the number of header lines per file?).
5. committed versions of notebooks on github/docs should be executed to show intermediate outputs/figures.
6. there should be a doc version of the notebook walkthroughs, cannot navigate to them from doc site.
7. more descriptive notebook names - preserve the order of 1>5, though, so lead names with those numbers like `N1_demand_modification.ipynb`.
8. where notebooks use external methods to fit distributional models, include brief notes about model structure and intuition behind it, visualize parameter learning if possible, and include any useful math as background. collapse this inline, or link to elsewhere as supplemental material. it is preferable to use minimized implementations with common libraries (numpy, scipy - see [here](https://github.com/ilaria-manco/hidden-markov-models/blob/master/HMM.ipynb)). though interesting, the conclusions at the end of notebook 4 about drought modeling are not explained.
