### 2024-03-09 IMMM-SFA/statemodify#55 @ekblad
sampling docs and unit tests
re: openjournals/joss-reviews#6325
a claim made here is that `statemodify` samples plausible futures based on input data. plausible in what sense? the two sampling methods used here, latin hypercube sampling and hidden markov modeling, need supplementary documentation and unit tests. demonstrate how specific aspects of time series distributions are preserved or modified by these methods (mean, variance, autocorrelation, trends). in the documentation, provide guidance on which method should be used for specific types of state space exploration. otherwise, the impact of using `lhs` vs. `hmm` in specific notebooks/modify functions is unclear.
