### 2024-03-09 #6325 @ekblad
hi all, i have finished with my comments for this version. in general, the library works as advertised, and documentation is nearing completion. some additional notes at IMMM-SFA/statemodify#54 can improve the user experience. a suggestion for two additional tests at IMMM-SFA/statemodify#55 provides some methodological confidence when using this library in applications other than the CRB. docs from `SALib` or elsewhere can be adapted for this purpose. neither issue is blocking, they are suggestions to improve uptake. thanks for the invite!
