### 2024-03-12 IMMM-SFA/statemodify#56 @rg727
when the user should see something, show it to them. example is when saying "In the input_files directory, you can open the sj2015B_S0_1.ddr file and see that the Admin # of our selected user has now become 1.0000." two ideas for this: 1a. use pandas immediately after to query the information that was described in words 1b. print directory contents or trees from cells using ! 
https://github.com/IMMM-SFA/statemodify/issues/54
https://github.com/openjournals/joss-reviews/issues/6325
