### 2024-03-12 IMMM-SFA/statemodify#58 @rg727
in structuring notebooks as the application of two or more major functions, minimize the surrounding code (do users need to cast several dictionary elements to float?).
@crvernon this has to do with the way that the .parquet files save everything as strings

[IMMM-SFA/statemodify#54](https://github.com/IMMM-SFA/statemodify/issues/54)
https://github.com/openjournals/joss-reviews/issues/6325
