### 2024-03-12 IMMM-SFA/statemodify#59 @rg727
collapsing config into a single .json or .yaml would save some lines (do users really need to set the number of header lines per file?).
https://github.com/IMMM-SFA/statemodify/issues/54
https://github.com/openjournals/joss-reviews/issues/6325
