### 2024-03-12 IMMM-SFA/statemodify#60 @rg727
committed versions of notebooks on github/docs should be executed to show intermediate outputs/figures.
https://github.com/IMMM-SFA/statemodify/issues/54
https://github.com/openjournals/joss-reviews/issues/6325
