### 2024-03-12 IMMM-SFA/statemodify#61 @rg727
there should be a doc version of the notebook walkthroughs, cannot navigate to them from doc site.
https://github.com/IMMM-SFA/statemodify/issues/54
https://github.com/openjournals/joss-reviews/issues/6325
