### 2024-03-12 IMMM-SFA/statemodify#62 @rg727
more descriptive notebook names - preserve the order of 1>5, though, so lead names with those numbers like N1_demand_modification.ipynb.
https://github.com/IMMM-SFA/statemodify/issues/54
https://github.com/openjournals/joss-reviews/issues/6325
