### 2024-03-12 IMMM-SFA/statemodify#63 @rg727
where notebooks use external methods to fit distributional models, include brief notes about model structure and intuition behind it, visualize parameter learning if possible, and include any useful math as background. collapse this inline, or link to elsewhere as supplemental material. it is preferable to use minimized implementations with common libraries (numpy, scipy - see here). though interesting, the conclusions at the end of notebook 4 about drought modeling are not explained.
https://github.com/IMMM-SFA/statemodify/issues/54
https://github.com/openjournals/joss-reviews/issues/6325
