### 2024-03-21 IMMM-SFA/statemodify#63 @rg727
Our focus in this notebook is to highlight the functionality of statemodify, not to highlight the HMM-based model specifically, but we agree with the reviewer that it’s important to point to supplemental information so that a user can learn more about HMM-based synthetic streamflow generation on their own. Thus, we have added a note box with links to useful resources that demonstrate the model fitting and validation. We used a standard library for fitting the HMM (hmmlearn). In regards to the last point on explanation of drought modeling, we provide explanations on the value of the hydrology generated and impacts to reservoirs which would support a more robust drought modeling experiment.
