### 2024-04-08 IMMM-SFA/statemodify#59 @crvernon
These were left as single files with purpose so that users would only have to interact with file types they had specific interests in.  We let the users set details like the number of header lines per file because these actually do change for some StateMod versions / implementations.  We wanted to give users the ability to customize these key things in case they were utilizing custom StateMod versions.  Thanks!
