### 2024-04-08 1cf7c79
limit numpy to <2 for pyarrow
