### 2024-04-10 ec54a41
add option for xdd dtype conversion
