### 2024-04-10 6744022
propagate call to dtypes to fn
