+++ docs/source/notebooks/N5_Batch_Modification.rst
@@ -1,0 +1,273 @@
+``statemodify`` Quickstarter Notebook #5 : Sampling Multiple Uncertainties
+--------------------------------------------------------------------------
+
+In the prior notebooks, we sampled just one type of uncertainty at a
+time to demonstrate how a single adjustment in the selected input file
+leads to a verifiable change in output so that we can demonstrate that
+``statemodify`` is working as expected. It is much harder to make sense
+of the relative importance of uncertain drivers unless we conduct a
+formal sensitivity analysis, which lies outside of the bounds of this
+tutorial. However, it is very likely that many of these uncertainties
+will be present and of interest in any given future for the region.
+Thus, this notebook is used to demonstrate how to do a Latin hypercube
+sample simultaneously across multiple uncertainties in a given basin
+using the ``modify_batch()`` function.
+
+.. code:: ipython3
+
+    import argparse
+    import logging
+    import os
+    import pickle
+    from string import Template
+    import subprocess
+    
+    import matplotlib.pyplot as plt
+    import numpy as np
+    import pandas as pd 
+    import statemodify as stm
+
+.. container:: alert alert-block alert-info
+
+   NOTE: Each simulation in this notebook is run for the length of the
+   historical period (from 1909-2013). If you want to reduce the length
+   of the simulation, navigate to the ``.ctl`` file and adjust the
+   ``iystr`` and ``iyend`` variables. For this notebook, this file is
+   located in: ``data/cm2015_StateMod/StateMod/cm2015.ctl``
+
+.. code:: ipython3
+
+    # statemod directory
+    statemod_dir = "/usr/src/statemodify/statemod_upper_co"
+    
+    # root directory of statemod data for the target basin
+    root_dir = os.path.join(statemod_dir, "src", "main", "fortran")
+    
+    # home directory of notebook instance
+    home_dir = os.path.dirname(os.getcwd())
+    
+    # path to the statemod executable
+    statemod_exe = os.path.join(root_dir, "statemod")
+    
+    # data directory and root name for the target basin
+    data_dir = os.path.join(
+        home_dir,
+        "data",
+        "cm2015_StateMod",
+        "StateMod"
+    )
+    
+    # directory to the target basin input files with root name for the basin
+    basin_path = os.path.join(data_dir, "cm2015B")
+    
+    # scenarios output directory
+    scenarios_dir = os.path.join(data_dir, "scenarios")
+    
+    #parquet output directory
+    parquet_dir=os.path.join(data_dir, "parquet")
+    
+    
+    # path to template file
+    multi_template_file = os.path.join(
+        home_dir,
+        "data",
+        "cm2015B_template_multi.rsp"
+    )
+
+Step 1: Creating a Sample Across Multiple Uncertainties
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+In this example, we create a global Latin hypercube sample across 3
+example uncertainties that we are interested in for the Upper Colorado:
+evaporation at reservoirs, modification of water rights, and demands.
+The form of the ``modify_batch()`` function is similiar to those
+presented in the last notebooks; however, now, a problem dictionary
+stores the names of the respective ``statemodify`` functions
+(``modify_eva``, ``modify_ddr``, ``modify_ddm``) that will now be
+applied simultaneously. The Latin hypercube sample is conducted with
+respect to the ``bounds`` listed and the resulting multipliers or
+additives are applied to the target IDs listed.
+
+.. container:: alert alert-block alert-info
+
+   NOTE: If you are interested in creating an alternative ``.ddr`` file
+   that does not sample decrees, only adjusts the water rights, then you
+   will need to write ``None`` in the “bounds” parameter as we do below.
[…]
