+++ docs/source/notebooks/N1_Demand_WaterRights_File_Modification.rst
@@ -1,0 +1,1055 @@
+``statemodify`` Quickstarter Notebook #1: Getting Started and Using the DDM and DDR Modification Functions in the San Juan River Basin
+--------------------------------------------------------------------------------------------------------------------------------------
+
+In this series of five notebooks, we demonstrate the functionality of
+``statemodify`` using three of the five subbasins on the West Slope
+basin in the state of Colorado: Gunnison, San Juan/Dolores, and the
+Upper Colorado Basin. There are two classes of adjustments offered in
+``statemodify`` that can be used to create alternative future states of
+the world for the region:
+
+1. Application of multipliers or additives to the original dataset which
+   are sampled from specified bounds using a Latin hypercube sample
+
+2. Complete swap of input data with data generated from an external
+   method
+
+Option 1 is applicable to ``.ddm`` (monthly demand), ``.ddr`` (water
+rights), ``.eva`` (reservoir evaporation), ``.res`` (reservoir storage).
+
+Option 2 is applicable to ``.xbm`` (monthly streamflow) and ``.iwr``
+(irrigation demand). In ``statemodify`` we provide a Hidden Markov Model
+(HMM)-based approach to generate synthetic flows across the basins and
+tie in irrigation demand to be negatively correlated to increased
+streamflow.
+
+In this first notebook, we will demonstrate now to use the demand
+(``modify_ddm()``)and water rights (``modify_ddr()``) modification
+functions in the San Juan River Basin. Demands are projected to increase
+with the growth of cities and agriculture and water rights will likely
+change as discussions on changes to the Colorado Compact and
+re-allocation of water across the Colorado River Basin to promote
+sustainable development continue.
+
+.. container:: alert alert-block alert-warning
+
+   Tip: When each StateMod file is mentioned, clicking on the name will
+   link the user to the StateMod documentation with more information on
+   that file.
+
+Step 1: Run a Historical Simulation in StateMod for the San Juan Basin
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Before we start on an exploratory modeling journey, you may be first
+interested in understanding water shortages that the basin has
+historically experienced. In the container, we have downloaded and
+compiled StateMod, ``statemodify``, and the San Juan dataset from the
+Colorado’s Decision Support System (CDSS) website. We can run a baseline
+simulation below which takes approximately 4 minutes. In this baseline
+simulation, we run StateMod over the length of the historical period
+(105 years) under the assumption that we are starting from current
+conditions.
+
+.. code:: ipython3
+
+    import argparse
+    import logging
+    import os
+    import pickle
+    from string import Template
+    import subprocess
+    
+    import matplotlib.pyplot as plt
+    import numpy as np
+    import pandas as pd 
+    import statemodify as stm
+
+.. container:: alert alert-block alert-info
+
+   NOTE: Each simulation in this notebook is run for the length of the
+   historical period (from 1909-2013). If you want to reduce the length
+   of the simulation, navigate to the ``.ctl`` file and adjust the
+   ``iystr`` and ``iyend`` variables. For this notebook, these files are
+   located in:
+   ``data/sj2015_StateMod_modified/sj2015_StateMod_modified/StateMod Notebook/sj2015.ctl``
+   and
[…]

[…]
