+++ docs/source/notebooks/N2_Evaporation_File_Modification.rst
@@ -25,10 +25,10 @@ dataset for the Gunnison.
     import pickle
     from string import Template
     import subprocess
-    
+
     import matplotlib.pyplot as plt
     import numpy as np
-    import pandas as pd 
+    import pandas as pd
     import statemodify as stm
 
 .. container:: alert alert-block alert-info
@@ -44,16 +44,16 @@ dataset for the Gunnison.
 
     # statemod directory
     statemod_dir = "/usr/src/statemodify/statemod_gunnison_sjd"
-    
+
     # root directory of statemod data for the target basin
     root_dir = os.path.join(statemod_dir, "src", "main", "fortran")
-    
+
     # home directory of notebook instance
     home_dir = os.path.dirname(os.getcwd())
-    
+
     # path to the statemod executable
     statemod_exe = os.path.join(root_dir, "statemod-17.0.3-gfortran-lin-64bit-o3")
-    
+
     # data directory and root name for the target basin
     data_dir = os.path.join(
         home_dir,
@@ -62,13 +62,13 @@ dataset for the Gunnison.
         "gm2015_StateMod_modified",
         "StateMod"
     )
-    
+
     # directory to the target basin input files with root name for the basin
     basin_path = os.path.join(data_dir, "gm2015B")
-    
+
     # scenarios output directory
     scenarios_dir = os.path.join(data_dir, "scenarios")
-    
+
     # path to eva template file
     eva_template_file = os.path.join(
         home_dir,
@@ -85,27 +85,27 @@ dataset for the Gunnison.
 
 .. parsed-literal::
 
-     Startup log file for messages to this point: /home/jovyan/data/gm2015_StateMod_modified/gm2015_StateMod_modified/StateMod/gm2015B.rsp                                                                                                                                                                        
+     Startup log file for messages to this point: /home/jovyan/data/gm2015_StateMod_modified/gm2015_StateMod_modified/StateMod/gm2015B.rsp
        Closing startup log file: statem.log
-       Opening dataset log file: /home/jovyan/data/gm2015_StateMod_modified/gm2015_StateMod_modified/StateMod/gm2015B.log                                                                                                                                                                        
+       Opening dataset log file: /home/jovyan/data/gm2015_StateMod_modified/gm2015_StateMod_modified/StateMod/gm2015B.log
     ________________________________________________________________________
-    
-            StateMod                       
-            State of Colorado - Water Supply Planning Model     
-    
-            Version:     17.0.3          
+
+            StateMod
+            State of Colorado - Water Supply Planning Model
+
+            Version:     17.0.3
             Last revision date: 2021/09/12
-    
+
     ________________________________________________________________________
-      
+
       Subroutine Execut
       Subroutine Datinp
-    
+
     ...
 
     ________________________________________________________________________
       Execut; Successful Termination
-      Statem; See detailed messages in dataset log file: /home/jovyan/data/gm2015_StateMod_modified/gm2015_StateMod_modified/StateMod/gm2015B.log                                                                                                                                                                        
+      Statem; See detailed messages in dataset log file: /home/jovyan/data/gm2015_StateMod_modified/gm2015_StateMod_modified/StateMod/gm2015B.log
      Stop 0
 
 
@@ -122,21 +122,21 @@ for supplying emergency water to Lake Powell.
 
 .. code:: ipython3
 
[…]
