### 2024-04-10 IMMM-SFA/statemodify#56 @crvernon
@rg727 addressed this by placing a documentation link into the notebook that points to the source info.
