### 2024-04-10 IMMM-SFA/statemodify#55 @rg727
The utility of statemodify is that it allows a concise and efficient way to sample plausible futures that can then be used to force StateMod, a water allocation model for the state of Colorado. A Hidden Markov Model (HMM)-based synthetic streamflow generator is fit on the observed historic streamflow traces in each subbasin of interest. Sampling from this model allows the user to create plausible synthetic traces that lie within the bounds of the historical distribution. Supplemental links that provide more information on the HMM model fitting and validation are now included in Notebook 4. The rest of the notebooks utilize a Latin hypercube sample to directly sample multipliers or additives of interest within some pre-specified bounds for each uncertainty (i.e. 0.5-1.5x for demands or an addition of -0.5-1ft/month loss in reservoir storage from evaporation). In order to make this sampling strategy more clear, we more carefully delineate where the notebooks that Latin hypercube sampling is occurring. We also more clearly state this at the beginning of the introductory notebook.
