### 2024-04-10 IMMM-SFA/statemodify#42 @crvernon
Addressed in #66
