+++ docs/build/html/_sources/getting-started/installation.rst
@@ -1,0 +1,17 @@
+Installation
+============
+
+Install ```statemodify``` using pip:
+
+.. code-block:: console
+
+  $ # statemodify requires python >= 3.8.
+  $ pip install statemodify
+
+You may also install ```statemodify``` from source.
+
+.. code-block:: console
+
+  $ git clone https://github.com/IMMM-SFA/statemodify.git
+  $ cd statemodify
+  $ pip install .
