+++ docs/build/html/_sources/index.rst
@@ -1,0 +1,129 @@
+Welcome to **statemodify**!
+--------------------------------
+
+Summary
+++++++++++++++++++++++++++++
+
+The Colorado River Basin (CRB) is experiencing an unprecedented water shortage crisis brought upon by a combination of factors arising from interactions across the region’s coupled human and natural systems. Allocation of water to the seven states that rely on the Colorado River was settled in the Colorado River Compact of 1922 during a period now known to be characterized by atypically high flow (Christensen et al., 2004). Since then, aridification due to anthropogenic-driven warming has steadily reduced the overall water supply available in the basin, with a 10% decrease in the river’s flow occurring over just the past two decades (Bass et al., 2023). The river is further strained by increasing demands associated with a growing population and diverse multi-sectoral demands. Navigating these challenges also requires accounting for the complex prior appropriation water rights system governing water allocation across the region’s diverse users.
+
+The state of Colorado’s West Slope basins are a critical component of the Colorado River System and reflect the broader challenges faced by the entire basin. The six West Slope basins – the Upper Colorado, Yampa, White, San Juan, Dolores, and Gunnison basins – comprise the headwaters of the Colorado River and contribute over 60% of the inflow to Lake Powell in an average year (Salehabi et al., 2020). The West Slope basins represent an essential part of the State of Colorado’s economy, supporting a multibillion-dollar tourism industry, providing water for roughly 24,000 acres of irrigated farmland, and sending drinking water across the continental divide to major metropolitan areas in eastern Colorado (State of Colorado, 2023).  Uncertainty stemming from climate change and institutional response plays a dominant role in evaluations of future deliveries to Lake Powell and characterization of the basin users’ vulnerabilities (Hadjimichael et al., 2020a; Hadjimichael et al., 2020b, Salehabi et al., 2022). Recent studies estimate that changes in temperature and precipitation may result in streamflows that are between 5% and 80% lower by the end of the 21st century when compared to the historical record (Kopytkovskiy et al., 2015; Milly and Dunn, 2020; Miller et al., 2021). Institutional responses to changes in flow, such as changes to reservoir operations and water rights structures, are difficult to predict and model using traditional probabilistic methods (Hadimichael et al., 2020). This difficulty in accurately characterizing key system inputs with known probability distributions is often described as conditions of "deep uncertainty" (Lempert, 2002 ; Kwakkel et al., 2016).
+
+To account for the deeply uncertain future conditions in the West Slope basins, approaches are needed that can help facilitate an understanding of vulnerabilities across many plausible future scenarios (Walker and Marchau, 2003; Lempert, 2002; Marchau et al., 2019). Exploratory modeling is one such approach that uses computational experiments to understand a range of possible model behaviors (Bankes, 1993). In the West Slope basins, exploratory modeling can be done with StateMod, a highly resolved, open source, regional water allocation model developed and maintained jointly by the Colorado Water Conservation Board (CWCB) and the Colorado Division of Water Resources (DWR) that is currently used to support water use assessments for the State of Colorado. The input files of StateMod can be manipulated to develop hypothetical scenarios to assess how changes in hydrology, water rights, or infrastructure impact regional water shortages, streamflow, or reservoir levels.
+
[…]
