### 2024-04-10 #6325 @rg727
Hi @cheginit, thank you for the ping. @crvernon and I have addressed the comments and made changes to our code and introductory notebooks. All responses are linked back to this thread and each comment is addressed in the closed issues (see here: https://github.com/IMMM-SFA/statemodify/issues?q=is%3Aissue+is%3Aclosed)
