### 2024-04-11 #6325 @barneydobson
@cheginit Happy with these revisions and all checkboxes ticked above.

@rg727 and @crvernon - in particular great job revising the notebooks, I think these changes will be a big QoL improvement for users!
