### 2024-04-24 #6325 @rg727
Hi @cheginit thanks for overseeing this process! Quick question: the publications missing DOIs are ones that don't have any DOIs associated with them. One is a state water plan, another is an older paper, and the final one is a New York Times article. Do you have any recommendations? Thanks!
