### 2024-04-24 #6325 @rg727
In response to the checklist

Version:  0.2.1
DOI for archive:  10.5281/zenodo.11061890

> * [ ]  Go through the missing DOIs listed above and add them.
> * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Make sure that the license listed for the archive is the same as the software license!
> * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [x]  Please list the DOI of the archived version here.
