### 2024-04-24 #6325 @cheginit
@rg727 Congratulations on your publication at JOSS!

While EiC/AEiC works on the final publication, please double-check the final PDF in the previous comment to ensure everything is correct, e.g., authors names and affiliations.
