### 2024-04-24 #6325 @kthyng
@rg727 We prefer to have the title and author list of the Zenodo archive associated with your JOSS submission exactly match the JOSS paper for consistency. Would you update the Zenodo metadata to do so?
