### 2024-04-24 #6325 @kthyng
@rg727 Also your paper looks good but please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
