### 2024-04-25 #6325 @kthyng
Congratulations on your new publication @rg727! Many thanks to @cheginit and to reviewers @barneydobson and @ekblad for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
